fix for multivalued query params on server side
