+++ modules/compliance-tests/src/smithy4s/compliancetests/internals/ClientHttpComplianceTestCase.scala
@@ -17,8 +17,6 @@
 package smithy4s.compliancetests
 package internals
 
-import java.nio.charset.StandardCharsets
-
 import cats.implicits._
 import org.http4s.headers.`Content-Type`
 import org.http4s.HttpApp
@@ -66,34 +64,12 @@ private[compliancetests] class ClientHttpComplianceTestCase[
       }
       .getOrElse(assert.success.pure[F])
 
-    def splitQuery(queryString: String): (String, String) = {
-      queryString.split("=", 2) match {
-        case Array(k, v) =>
-          (
-            k,
-            Uri.decode(
-              toDecode = v,
-              charset = StandardCharsets.UTF_8,
-              plusIsSpace = true
-            )
-          )
-        case Array(k) => (k, "")
-      }
-    }
     val expectedUri = baseUri
       .withPath(
         Uri.Path.unsafeFromString(testCase.uri)
       )
       .withMultiValueQueryParams(
-        testCase.queryParams.combineAll
-          .map(splitQuery)
-          .foldRight[Map[String, List[String]]](Map.empty) {
-            case ((k, v), acc) =>
-              acc.get(k) match {
-                case Some(value) => acc + (k -> (v :: value))
-                case None        => acc + (k -> List(v))
-              }
-          }
+        parseQueryParams(testCase.queryParams)
       )
 
     val uriAssert = assert.eql(expectedUri, request.uri)
